+++ apps/admin/package.json
@@ -19,6 +19,7 @@
     "@hookform/resolvers": "^5.2.2",
     "@hugeicons/core-free-icons": "^4.0.0",
     "@hugeicons/react": "^1.1.6",
+    "@internationalized/date": "^3.12.0",
     "@tailwindcss/vite": "^4.2.1",
     "@tanstack/react-query": "catalog:",
     "clsx": "^2.1.1",
