+++ apps/admin/src/entities/cohort/index.ts
@@ -3,5 +3,4 @@ export * from "./model/statusFlow"
 export * from "./model/completion"
 export * from "./model/serialize"
 export * from "./model/useCreateOrUpdateCohortFlow"
-export * from "./model/useDeleteCohortFlow"
 export * from "./model/useTransitionCohortStatusFlow"
