+++ apps/admin/src/pages/semesters/SemestersPage.tsx
@@ -2,14 +2,11 @@ import { useMemo, useState } from "react"
 import { PlusSignIcon } from "@hugeicons/core-free-icons"
 import { HugeiconsIcon } from "@hugeicons/react"
 
-import { Button, Input, ListBox, Select, Table } from "@heroui/react"
+import { Button } from "@heroui/react"
 
 import type { CohortDto } from "@ddd/api"
 
 import {
-  STATUS_LABEL,
-  NEXT_STATUS_BUTTON_LABEL,
-  nextStatus,
   serializeCohortToForm,
   useTransitionCohortStatusFlow,
 } from "@/entities/cohort"
@@ -18,15 +15,10 @@ import { GridBox } from "@/shared/ui/GridBox"
 import { StatCard } from "@/shared/ui/StatCard"
 import { Description, Title } from "@/widgets/heading"
 
-import { SemesterRegisterDrawer } from "./components"
-import {
-  STATUS_FILTER_OPTIONS,
-  type StatusFilterValue,
-} from "./constants"
+import { SemesterRegisterDrawer, SemesterTableSection } from "./components"
 import {
   useSemesterRegistrationMode,
   useSemestersTableData,
-  type CohortRow,
   type SemestersSummary,
 } from "./hooks"
 
@@ -35,9 +27,6 @@ export default function SemestersPage() {
   const { tableRows, summary, isError, refetch } = useSemestersTableData()
   const registration = useSemesterRegistrationMode()
   const { transition } = useTransitionCohortStatusFlow()
-
-  const [searchText, setSearchText] = useState("")
-  const [statusFilter, setStatusFilter] = useState<StatusFilterValue>("ALL")
 
   // 행 "수정" 클릭 시 채워지는 edit 타겟. registration 모드를 오버라이드.
   const [editTarget, setEditTarget] = useState<CohortDto | null>(null)
@@ -62,16 +51,6 @@ export default function SemestersPage() {
     setIsDrawerOpen(open)
     if (!open) setEditTarget(null)
   }
-
-  const filteredRows = useMemo(() => {
-    return tableRows
-      .filter((row) =>
-        searchText === "" ? true : row.name.includes(searchText),
-      )
-      .filter((row) =>
-        statusFilter === "ALL" ? true : row.status === statusFilter,
-      )
-  }, [tableRows, searchText, statusFilter])
 
   if (isError) {
     return (
@@ -108,77 +87,15 @@ export default function SemestersPage() {
         prefill={drawerProps.prefill}
       />
 
-      <div className="space-y-5 rounded-lg bg-white p-5 shadow">
-        <FlexBox className="justify-between">
-          <Input
-            variant="secondary"
-            placeholder="검색..."
-            className="max-w-xs"
-            value={searchText}
-            onChange={(e) => setSearchText(e.target.value)}
-          />
-          <Select
-            variant="secondary"
-            className="max-w-36"
-            aria-label="상태 필터"
-            selectedKey={statusFilter}
-            onSelectionChange={(key) =>
-              setStatusFilter(String(key) as StatusFilterValue)
-            }
-          >
-            <Select.Trigger>
-              <Select.Value>
-                {
-                  STATUS_FILTER_OPTIONS.find((o) => o.value === statusFilter)
-                    ?.label
-                }
-              </Select.Value>
-              <Select.Indicator />
-            </Select.Trigger>
-            <Select.Popover>
-              <ListBox>
-                {STATUS_FILTER_OPTIONS.map((option) => (
-                  <ListBox.Item
-                    key={option.value}
-                    id={option.value}
-                    textValue={option.label}
-                  >
-                    {option.label}
-                  </ListBox.Item>
-                ))}
-              </ListBox>
-            </Select.Popover>
-          </Select>
-        </FlexBox>
-
-        <Table>
-          <Table.ScrollContainer>
-            <Table.Content aria-label="기수 목록">
-              <Table.Header>
-                <Table.Column isRowHeader>기수</Table.Column>
-                <Table.Column>상태</Table.Column>
-                <Table.Column>모집 기간</Table.Column>
-                <Table.Column>지원자 수</Table.Column>
-                <Table.Column>멤버 수</Table.Column>
-                <Table.Column>등록일</Table.Column>
-                <Table.Column>액션</Table.Column>
-              </Table.Header>
-              <Table.Body>
-                {filteredRows.map((row) => (
-                  <SemesterRow
-                    key={row.id}
-                    row={row}
-                    onEdit={() => {
-                      setEditTarget(row)
-                      setIsDrawerOpen(true)
-                    }}
-                    onTransition={() => transition(row)}
-                  />
-                ))}
-              </Table.Body>
-            </Table.Content>
-          </Table.ScrollContainer>
-        </Table>
+      <div className="rounded-lg bg-white p-5 shadow">
+        <SemesterTableSection
+          rows={tableRows}
+          onEditRow={(row) => {
+            setEditTarget(row)
+            setIsDrawerOpen(true)
+          }}
+          onTransitionRow={(row) => transition(row)}
+        />
       </div>
     </div>
   )
@@ -238,47 +155,4 @@ const CardSection = ({ summary }: CardSectionProps) => {
       />
     </GridBox>
   )
-}
-
-type SemesterRowProps = {
-  row: CohortRow
-  onEdit: () => void
-  onTransition: () => void
-}
-
-const SemesterRow = ({ row, onEdit, onTransition }: SemesterRowProps) => {
-  const transitionLabel = NEXT_STATUS_BUTTON_LABEL[row.status]
-  const canTransition = nextStatus(row.status) !== null
-
-  return (
-    <Table.Row>
-      <Table.Cell>{row.name}</Table.Cell>
-      <Table.Cell>{STATUS_LABEL[row.status]}</Table.Cell>
-      <Table.Cell>
-        {formatPeriod(row.recruitStartAt, row.recruitEndAt)}
-      </Table.Cell>
-      <Table.Cell>{row.applicantsCount ?? "-"}</Table.Cell>
-      <Table.Cell>{row.membersCount ?? "-"}</Table.Cell>
-      <Table.Cell>
-        {new Date(row.createdAt).toLocaleDateString("ko-KR")}
-      </Table.Cell>
-      <Table.Cell>
-        <Button size="sm" variant="outline" className="mr-2" onPress={onEdit}>
-          수정
-        </Button>
-        {canTransition && transitionLabel && (
-          <Button size="sm" onPress={onTransition}>
-            {transitionLabel}
-          </Button>
-        )}
-      </Table.Cell>
-    </Table.Row>
-  )
-}
-
-const formatPeriod = (start: string, end: string): string => {
-  if (!start && !end) return "-"
-  const left = start ? start.slice(0, 10) : "?"
-  const right = end ? end.slice(0, 10) : "?"
-  return `${left} ~ ${right}`
 }
