+++ apps/admin/src/pages/semesters/components/index.ts
@@ -1,2 +1,3 @@
 export { SemesterRegisterDrawer } from "./SemesterRegisterDrawer"
 export type { DrawerMode } from "./SemesterRegisterDrawer"
+export { SemesterTableSection } from "./SemesterTableSection"
